Limit number of objects loaded, show map message
Endpoint functions check the number of objects in the requested area and
abort the request early with 413 response if necessary. DataSource
refresh method now returns a promise and error handling is moved to
index.js. When 413 response is received, the Leaflet message control
displays a message to the user.

Related to #1